Move requestTokenIntrospect() to LoginService
